Fixes Ticket #1476:  * Uses String.Empty rather than ""
Note: The advantage of String.Empty is that it shows you meant to use an empty string rather than just typing "" and forgetting to put anything in there. Some people will agree with this methodology, some will disagree I am sure. For me personally I have a G15 key configured to type String.Empty when I need it, so it is actually fewer keystrokes.